Implemented just for CPU and Metal.
With `Q4_0` there is ~8% performance hit for TG 500 and ~4% for PP 512, but might be possible to optimize the rope kernel further to compensate