@slaren 2023-09-18
These are the results I obtained on CPU (13900k):

| model                          |       size |     params | backend    |    th| test       |       Master t/s |           PR t/s | speedup         |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ---------- | ---------------: | ---------------: | --------------: |
| LLaMA 7B mostly Q4_0           |   3.56 GiB |     6.74 B | CPU        |          8 | pp 512     |     34.35 ± 0.31 |     36.52 ± 0.70 | 1.06   |
| LLaMA 7B mostly Q4_0           |   3.56 GiB |     6.74 B | CPU        |          8 | tg 128     |     16.03 ± 0.21 |     15.59 ± 0.16 | 0.97   |
| LLaMA 7B mostly Q4_0           |   3.56 GiB |     6.74 B | CPU        |          8 | tg 256     |     15.69 ± 0.35 |     14.97 ± 0.28 | 0.95   |
| LLaMA 7B mostly Q4_0           |   3.56 GiB |     6.74 B | CPU        |          8 | tg 512     |     15.56 ± 1.06 |     13.87 ± 0.06 | 0.89   |
| LLaMA 7B mostly Q4_0           |   3.56 GiB |     6.74 B | CPU        |          8 | tg 1024    |     15.12 ± 0.17 |     11.44 ± 0.22 | 0.76   |
| LLaMA 7B mostly Q4_0           |   3.56 GiB |     6.74 B | CPU        |          8 | tg 2048    |     14.60 ± 0.22 |      8.48 ± 0.10 | 0.58    |
| LLaMA 7B mostly Q4_0           |   3.56 GiB |     6.74 B | CPU        |          8 | tg 4096    |     12.89 ± 0.10 |      5.52 ± 0.04 | 0.42    |
| LLaMA 7B mostly Q4_0           |   3.56 GiB |     6.74 B | CPU        |          8 | tg 8192    |     10.55 ± 0.11 |      3.25 ± 0.01 | 0.31    |