@ggerganov 2023-09-18
Yup, these results might be a strong argument against the non-RoPEd K cache.