@slaren 2023-09-18
I suspect that the main cost is the copy of `K`, but the computational cost of calculating the RoPE shouldn't be too bad, and most of it could be replaced with a lookup table if needed. So a fused attention op that applies RoPE on the fly during the computation of `KQ` could be a viable way to do this.