@ggerganov 2023-09-18
> I suspect that the main cost is the copy of K

This is likely the case. We can verify this by replacing the rope with a cpy.