@slaren 2023-10-03
You can test the best scenario of this by removing the mul mats with `wk`, `wq` and `wv`. I tested this, and with 7B models on CPU I got between 10% and 40% higher t/s, depending on the model. GQA models (mistral) benefit less. But I agree with @ggerganov that this wouldn't work for anything other than the first layer, and in that case the performance difference would very likely be negligible.